Use `size_t` for parameter and return value of CMSG_{LEN,SPACE}
Fixes: #3240

Matches all platforms except Solaris and its descendants. `size_t` is
used there regardless in order to force excessive conversions, and to
make cross-platform code easier to write.

[illumos]: https://github.com/illumos/illumos-gate/blob/118b2dbf1f4a745a7e35a5054a777c09bd90fff7/usr/src/uts/common/sys/socket.h#L499-L505

+++ src/new/common/solarish.rs
@@ -12,8 +12,8 @@ pub(crate) mod sys {
             crate::new::common::posix::sys::socket::align_impl(p, size_of::<c_int>())
         }
 
-        pub const fn CMSG_LEN(length: c_uint) -> c_uint {
-            _CMSG_DATA_ALIGN(size_of::<cmsghdr>()) as c_uint + length
+        pub const fn CMSG_LEN(length: size_t) -> size_t {
+            _CMSG_DATA_ALIGN(size_of::<cmsghdr>()) + length
         }
 
         pub unsafe fn CMSG_DATA(cmsg: *const cmsghdr) -> *mut c_uchar {
